+++ packages/app/src/cli/commands/app/connect.ts
@@ -1,7 +1,0 @@
-import {Command} from '@oclif/core'
-
-export default class Connect extends Command {
-  static description = 'Connect the local app to an app on Shopify'
-
-  async run(): Promise<void> {}
-}