+++ packages/cli-kit/src/api/graphql/create_app.ts
@@ -48,7 +48,7 @@ export interface CreateAppQuerySchema {
       redirectUrlWhitelist: string[]
       apiSecretKeys: {
         secret: string
-      }
+      }[]
       appType: string
     }
     userErrors: {