+++ fixtures/app/CHANGELOG.md
@@ -1,5 +1,28 @@
 # fixtures-app
 
+## 0.1.4
+
+### Patch Changes
+
+- Updated dependencies
+  - @shopify/app@1.0.6
+  - @shopify/cli@1.0.6
+
+## 0.1.3
+
+### Patch Changes
+
+- @shopify/app@1.0.5
+- @shopify/cli@1.0.5
+
+## 0.1.2
+
+### Patch Changes
+
+- Updated dependencies
+  - @shopify/app@1.0.4
+  - @shopify/cli@1.0.4
+
 ## 0.1.1
 
 ### Patch Changes