+++ packages/cli-hydrogen/CHANGELOG.md
@@ -1,5 +1,27 @@
 # @shopify/cli-hydrogen
 
+## 1.0.6
+
+### Patch Changes
+
+- Updated dependencies
+  - @shopify/cli-kit@1.0.6
+
+## 1.0.5
+
+### Patch Changes
+
+- Updated dependencies
+  - @shopify/cli-kit@1.0.5
+
+## 1.0.4
+
+### Patch Changes
+
+- e2e working
+- Updated dependencies
+  - @shopify/cli-kit@1.0.4
+
 ## 1.0.1
 
 ### Patch Changes