+++ packages/cli-kit/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @shopify/cli-kit
 
+## 1.0.6
+
+### Patch Changes
+
+- Add deploy command
+
+## 1.0.5
+
+### Patch Changes
+
+- Import ngrok dynamically
+
+## 1.0.4
+
+### Patch Changes
+
+- e2e working
+
 ## 1.0.1
 
 ### Patch Changes