+++ packages/create-hydrogen/CHANGELOG.md
@@ -1,5 +1,31 @@
 # @shopify/create-hydrogen
 
+## 1.0.6
+
+### Patch Changes
+
+- Add deploy command
+- Updated dependencies
+  - @shopify/cli-kit@1.0.6
+  - @shopify/cli-hydrogen@1.0.6
+
+## 1.0.5
+
+### Patch Changes
+
+- Updated dependencies
+  - @shopify/cli-kit@1.0.5
+  - @shopify/cli-hydrogen@1.0.5
+
+## 1.0.4
+
+### Patch Changes
+
+- e2e working
+- Updated dependencies
+  - @shopify/cli-kit@1.0.4
+  - @shopify/cli-hydrogen@1.0.4
+
 ## 1.0.1
 
 ### Patch Changes